Use cobra for command line arguments parsing
