+++ main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/lbryio/lbrytv/cmd"
+)
+
+func main() {
+	cmd.Execute()
+}
